+++ src/FluentDOM/NamespaceResolver.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace FluentDOM {
+
+  interface NamespaceResolver {
+
+    /**
+     * @param string $prefix
+     * @return string
+     */
+    function resolveNamespace($prefix);
+  }
+}